locode-db: move generate to main file
Move all from generate.go to main.go and make Execute main function.
